#304 refactor stone module config setting
