+++ src/stone/stone_config.py
@@ -1,0 +1,126 @@
+from src.stone.examples.stone_env import src_stone_dir
+
+
+def column_order_str() -> str:
+    return "column_order"
+
+
+def sort_order_str() -> str:
+    return "sort_order"
+
+
+def atom_categorys_str() -> str:
+    return "atom_categorys"
+
+
+def attributes_str() -> str:
+    return "attributes"
+
+
+def must_be_roadnode_str() -> str:
+    return "must_be_RoadNode"
+
+
+def must_be_roadunit_str() -> str:
+    return "must_be_RoadUnit"
+
+
+def must_be_str() -> str:
+    return "must_be_str"
+
+
+def must_be_number_str() -> str:
+    return "must_be_number"
+
+
+def must_be_bool_str() -> str:
+    return "must_be_bool"
+
+
+def get_stone_formats_dir() -> str:
+    return f"{src_stone_dir()}/stone_formats"
+
+
+def stone_format_00001_acct_v0_0_0() -> str:
+    return "stone_format_00001_acct_v0_0_0"
+
+
+def stone_format_00002_membership_v0_0_0() -> str:
+    return "stone_format_00002_membership_v0_0_0"
+
+
+def stone_format_00003_ideaunit_v0_0_0() -> str:
+    return "stone_format_00003_ideaunit_v0_0_0"
+
+
+def stone_format_00019_ideaunit_v0_0_0() -> str:
+    return "stone_format_00019_ideaunit_v0_0_0"
+
+
+# def stone_format_00020_bud_acct_membership_v0_0_0()-> str: return "stone_format_00020_bud_acct_membership_v0_0_0"
+# def stone_format_00021_bud_acctunit_v0_0_0()-> str: return "stone_format_00021_bud_acctunit_v0_0_0"
+# def stone_format_00022_bud_idea_awardlink_v0_0_0()-> str: return "stone_format_00022_bud_idea_awardlink_v0_0_0"
+# def stone_format_00023_bud_idea_factunit_v0_0_0()-> str: return "stone_format_00023_bud_idea_factunit_v0_0_0"
+# def stone_format_00024_bud_idea_grouphold_v0_0_0()-> str: return "stone_format_00024_bud_idea_grouphold_v0_0_0"
+# def stone_format_00025_bud_idea_healerhold_v0_0_0()-> str: return "stone_format_00025_bud_idea_healerhold_v0_0_0"
+# def stone_format_00026_bud_idea_reason_premiseunit_v0_0_0()-> str: return "stone_format_00026_bud_idea_reason_premiseunit_v0_0_0"
+# def stone_format_00027_bud_idea_reasonunit_v0_0_0()-> str: return "stone_format_00027_bud_idea_reasonunit_v0_0_0"
+# def stone_format_00028_bud_ideaunit_v0_0_0()-> str: return "stone_format_00028_bud_ideaunit_v0_0_0"
+# def stone_format_00029_budunit_v0_0_0()-> str: return "stone_format_00029_budunit_v0_0_0"
+def stone_format_00020_bud_acct_membership_v0_0_0() -> str:
+    return "stone_format_00020_bud_acct_membership_v0_0_0"
+
+
+def stone_format_00021_bud_acctunit_v0_0_0() -> str:
+    return "stone_format_00021_bud_acctunit_v0_0_0"
+
+
+def stone_format_00022_bud_idea_awardlink_v0_0_0() -> str:
+    return "stone_format_00022_bud_idea_awardlink_v0_0_0"
+
+
+def stone_format_00023_bud_idea_factunit_v0_0_0() -> str:
+    return "stone_format_00023_bud_idea_factunit_v0_0_0"
+
+
+def stone_format_00024_bud_idea_grouphold_v0_0_0() -> str:
+    return "stone_format_00024_bud_idea_grouphold_v0_0_0"
+
+
+def stone_format_00025_bud_idea_healerhold_v0_0_0() -> str:
+    return "stone_format_00025_bud_idea_healerhold_v0_0_0"
+
+
+def stone_format_00026_bud_idea_reason_premiseunit_v0_0_0() -> str:
+    return "stone_format_00026_bud_idea_reason_premiseunit_v0_0_0"
+
+
+def stone_format_00027_bud_idea_reasonunit_v0_0_0() -> str:
+    return "stone_format_00027_bud_idea_reasonunit_v0_0_0"
+
+
+def stone_format_00028_bud_ideaunit_v0_0_0() -> str:
+    return "stone_format_00028_bud_ideaunit_v0_0_0"
+
+
+def stone_format_00029_budunit_v0_0_0() -> str:
+    return "stone_format_00029_budunit_v0_0_0"
+
+
+def get_stone_filenames() -> set[str]:
+    return {
+        stone_format_00001_acct_v0_0_0(),
+        stone_format_00002_membership_v0_0_0(),
+        stone_format_00003_ideaunit_v0_0_0(),
+        stone_format_00019_ideaunit_v0_0_0(),
+        stone_format_00020_bud_acct_membership_v0_0_0(),
+        stone_format_00021_bud_acctunit_v0_0_0(),
+        stone_format_00022_bud_idea_awardlink_v0_0_0(),
+        stone_format_00023_bud_idea_factunit_v0_0_0(),
+        stone_format_00024_bud_idea_grouphold_v0_0_0(),
+        stone_format_00025_bud_idea_healerhold_v0_0_0(),
+        stone_format_00026_bud_idea_reason_premiseunit_v0_0_0(),
+        stone_format_00027_bud_idea_reasonunit_v0_0_0(),
+        stone_format_00028_bud_ideaunit_v0_0_0(),
+        stone_format_00029_budunit_v0_0_0(),
+    }
