+++ src/stone/test_stone/test_stone_format_config.py
@@ -23,16 +23,14 @@ from src.gift.atom_config import (
     gogo_want_str,
     stop_want_str,
 )
-from src.stone.stone import (
+from src.stone.stone import _generate_stone_dataframe, get_stoneref, _get_headers_list
+from src.stone.stone_config import (
     get_stone_formats_dir,
     get_stone_filenames,
     stone_format_00001_acct_v0_0_0,
     stone_format_00002_membership_v0_0_0,
     stone_format_00003_ideaunit_v0_0_0,
     stone_format_00019_ideaunit_v0_0_0,
-    _get_headers_list,
-    _generate_stone_dataframe,
-    get_stoneref,
 )
 from src.stone.examples.stone_env import src_stone_dir
