+++ src/stone/test_stone/test_stone_format_import.py
@@ -1,36 +1,31 @@
-from src._instrument.file import delete_dir, open_file, create_file_path as f_path
+from src._instrument.file import create_file_path as f_path
 from src._road.jaar_refer import sue_str, bob_str, yao_str
-from src.bud.idea import ideaunit_shop
 from src.bud.bud import budunit_shop
-from src.listen.hubunit import hubunit_shop
-from src.stone.stone import (
-    stone_format_00001_acct_v0_0_0,
-    stone_format_00002_membership_v0_0_0,
-    stone_format_00003_ideaunit_v0_0_0,
-    create_stone_df,
+from src.gift.atom_config import (
     real_id_str,
     owner_id_str,
     acct_id_str,
-    group_id_str,
-    parent_road_str,
-    label_str,
-    mass_str,
-    pledge_str,
     debtit_score_str,
     credit_score_str,
-    debtit_vote_str,
-    credit_vote_str,
+)
+from src.listen.hubunit import hubunit_shop
+from src.stone.stone import (
+    create_stone_df,
     get_stoneref,
     save_stone_csv,
     open_stone_csv,
     load_stone_csv,
+)
+from src.stone.stone_config import (
+    stone_format_00001_acct_v0_0_0,
+    stone_format_00002_membership_v0_0_0,
+    stone_format_00003_ideaunit_v0_0_0,
 )
 from src.stone.examples.stone_env import (
     stone_examples_dir,
     stone_reals_dir,
     stone_env_setup_cleanup,
 )
-from os.path import exists as os_path_exists
 
 
 def test_open_stone_csv_ReturnsObj():
